requirements.txt
requests==2.27.1
selenium==4.2.0
webdriver-manager==3.7.0
python-telegram-bot
ipython
